New Pull Request

+++ api/views.py
@@ -4,8 +4,6 @@ from rest_framework.response import Response
 from .serializers import NoteSerializer
 from .models import Note
 
-# Create your views here.
-
 @api_view(['GET'])
 def getRoutes(request):
     routes = [
@@ -44,35 +42,51 @@ def getRoutes(request):
 
 @api_view(['GET'])
 def getNotes(request):
-    notes = Note.objects.all().order_by('-created')
-    serializer = NoteSerializer(notes, many=True)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        notes = Note.objects.all().order_by('-created')
+        serializer = NoteSerializer(notes, many=True)
+        return Response(serializer.data)
 
 @api_view(['GET'])
 def getNote(request, pk):
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(note, many=False)
-    return Response(serializer.data)
+    if request.method == 'GET':
+        try:
+            note = Note.objects.get(id=pk)
+            serializer = NoteSerializer(note, many=False)
+            return Response(serializer.data)
+        except Note.DoesNotExist:
+            return Response(status=404)
 
 @api_view(['PUT'])
 def updateNote(request, pk):
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(instance=note, data=request.data)
-    if serializer.is_valid():
-        serializer.save()
-    return Response(serializer.data)
+    if request.method == 'PUT':
+        try:
+            note = Note.objects.get(id=pk)
+            serializer = NoteSerializer(instance=note, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(serializer.errors, status=400)
+        except Note.DoesNotExist:
+            return Response(status=404)
 
 @api_view(['DELETE'])
 def deleteNote(request, pk):
-    note = Note.objects.get(id=pk)
-    note.delete()
-    return Response('Note was deleted!')
+    if request.method == 'DELETE':
+        try:
+            note = Note.objects.get(id=pk)
+            note.delete()
+            return Response('Note was deleted!')
+        except Note.DoesNotExist:
+            return Response(status=404)
 
 @api_view(['POST'])
 def createNote(request):
-    data = request.data
-    note = Note.objects.create(
-        body=data['body']
-    )
-    serializer = NoteSerializer(note, many=False)
-    return Response(serializer.data)
+    if request.method == 'POST':
+        serializer = NoteSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=201)
+        else:
+            return Response(serializer.errors, status=400)
